deconvolution the displacement as well, instead of integration from the acceleration.
In real earthquake, the observed/measured displacement has extra information,
the results are different from the integration of the acceleration.
Deconvolution of disp directly is correct.

For the artificial wave (like ricker wave, has analytical expression), it
is fine. Integration and deconvolution to displacement have the same results.

The problem in sine wave test is the initial velocity/displacement
are not zero. Use the envolope of sine-wave instead to make sure
it has both zero initial vel and disp.

updated examples and readme.